requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.2
firebase-admin==4.0.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
multidict==4.7.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
